Extract Slack custom API runtime
