+++ internal/findings/coverage_control_domain_refs.yaml
@@ -28,6 +28,15 @@ control_domains:
     - { framework_name: "ISO 27001:2022", control_id: "A.5.15" }
     - { framework_name: "ISO 27001:2022", control_id: "A.5.17" }
     - { framework_name: "ISO 27001:2022", control_id: "A.5.18" }
+  access_review:
+    - { framework_name: "SOC 2", control_id: "CC6.2" }
+    - { framework_name: "SOC 2", control_id: "CC6.3" }
+    - { framework_name: "NIST 800-53 r5", control_id: "AC-2" }
+    - { framework_name: "NIST 800-53 r5", control_id: "AC-6" }
+    - { framework_name: "CIS Controls v8", control_id: "5" }
+    - { framework_name: "CIS Controls v8", control_id: "6" }
+    - { framework_name: "PCI DSS v4.0.1", control_id: "7.2" }
+    - { framework_name: "ISO 27001:2022", control_id: "A.5.18" }
   asset_inventory:
     - { framework_name: "SOC 2", control_id: "CC6.1" }
     - { framework_name: "SOC 2", control_id: "CC7.1" }
